Silence a couple of compiler warnings.
git-svn-id: http://svn.mythtv.org/svn/trunk@23101 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythfrontend/playbackbox.cpp
@@ -988,8 +988,8 @@ void PlaybackBox::UpdateUsageUI(void)
 
     if (usedProgress)
     {
-        usedProgress->SetTotal(freeSpaceTotal);
-        usedProgress->SetUsed(freeSpaceUsed);
+        usedProgress->SetTotal((int)freeSpaceTotal);
+        usedProgress->SetUsed((int)freeSpaceUsed);
     }
 }
